+++ dataprep-dataset/src/main/java/org/talend/dataprep/dataset/store/metadata/file/FileSystemDataSetMetadataRepository.java
@@ -70,11 +70,6 @@ public class FileSystemDataSetMetadataRepository extends DataSetMetadataReposito
             return null;
         }
 
-        if (!file.canRead()) {
-            LOG.info("dataset #{} not available in file system, it is perhaps used by another thread ?", id);
-            return null;
-        }
-
         try (ObjectInputStream input = new ObjectInputStream(new FileInputStream(file))) {
             return (DataSetMetadata) input.readObject();
         } catch (ClassNotFoundException | IOException e) {
@@ -88,7 +83,7 @@ public class FileSystemDataSetMetadataRepository extends DataSetMetadataReposito
     @Override
     public void remove(String id) {
         final File file = getFile(id);
-        if (file.exists()) {
+        if (file.exists() && !isNFSSpecificFile(file)) {
             file.delete();
         }
         LOG.debug("metadata {} successfully deleted", id);
@@ -157,6 +152,16 @@ public class FileSystemDataSetMetadataRepository extends DataSetMetadataReposito
      */
     private File getRootFolder() {
         return new File(storeLocation + "/metadata/");
+    }
+
+    /**
+     * Specific OS NFS files must be not be deleted.
+     * 
+     * @param file the file to read.
+     * @return True if the given belongs to the operating system and is NFS related (starts with ".nfs")
+     */
+    private boolean isNFSSpecificFile(File file) {
+        return file.getName().startsWith(".nfs");
     }
 
 }
